+++ client/galaxy/scripts/mvc/ui/ui-select-default.js
@@ -208,7 +208,7 @@ var View = Backbone.View.extend({
                         ${_.reduce(
                             filteredTags.slice(0, 5),
                             (memo, tag) => {
-                                return `${memo}&nbsp;<div class="badge badge-primary badge-tags">${_.escape(
+                                return `${memo}&nbsp;<div style="background-color: ${Utils.generateTagColor(tag.slice(5))}" class="badge badge-primary badge-tags">${_.escape(
                                     tag
                                 )}</div>`;
                             },
