+++ client/galaxy/scripts/utils/utils.js
@@ -349,6 +349,40 @@ export function setWindowTitle(title) {
     }
 }
 
+/**
+ * Calculate a 32 bit FNV-1a hash
+ * Found here: https://gist.github.com/vaiorabbit/5657561
+ * Ref.: http://isthe.com/chongo/tech/comp/fnv/
+ *
+ * @param {string} str the input value
+ * @returns {integer}
+ */
+function hashFnv32a(str) {
+    /*jshint bitwise:false */
+    var i, l,
+        hval = 0x811c9dc5;
+
+    for (i = 0, l = str.length; i < l; i++) {
+        hval ^= str.charCodeAt(i);
+        hval += (hval << 1) + (hval << 4) + (hval << 7) + (hval << 8) + (hval << 24);
+    }
+    return hval >>> 0;
+}
+
+export function generateTagColor(tagValue) {
+    //var r = (15 - ((hash & 0xf) % 6)).toString(16),
+        //g = (15 - (((hash & 0xf0) >> 4) % 6)).toString(16),
+        //b = (15 - (((hash & 0xf00) >> 8) % 6)).toString(16);
+
+    var hash = hashFnv32a(tagValue);
+    var r = (((hash & 0x00f) >> 0) % 10),
+        g = (((hash & 0x0f0) >> 4) % 10),
+        b = (((hash & 0xf00) >> 8) % 10);
+
+    return `#${r}${g}${b}`;
+}
+
+
 export default {
     cssLoadFile: cssLoadFile,
     cssGetAttribute: cssGetAttribute,
@@ -368,5 +402,6 @@ export default {
     linkify: linkify,
     appendScriptStyle: appendScriptStyle,
     getQueryString: getQueryString,
-    setWindowTitle: setWindowTitle
+    setWindowTitle: setWindowTitle,
+    generateTagColor: generateTagColor
 };
